requirements.txt
Flask==1.1.4
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Jinja2==2.11.3

Werkzeug==1.0.1

psycopg2==2.9.5
Flask-Migrate==3.0.0

Bcrypt-Flask==1.0.2
WTForms==2.3.3
email-validator==1.3.1

Flask-WTF==1.1.1

click==7.1.2

Flask-Admin==1.6.1

Flask-COMBO-JSONAPI==1.0.7
ComboJSONAPI==1.1.1
apispec==4.7.1
PyYAML==5.4.1

markupsafe==2.0.1
gunicorn==20.1.0
